fix(security): hardening phase 2 — resolve 9 audit findings (#1242)
fix(security): hardening phase 2 — resolve 9 audit findings (HIGH + MED)

On-chain (Rust):
- HIGH-7: active_tasks u8 → u16 (absorb byte from _reserved, SIZE unchanged)
- HIGH-6: multisig threshold >= 2 (multisig.rs, initialize_protocol, update_multisig)
- HIGH-9: slash_percentage <= 100 validation + compile-time assert
- MED-14: max deadline 1 year (MAX_DEADLINE_SECONDS in task_init_helpers)
- MED-15: min reputation > 0 for dispute voters

SDK (TypeScript):
- HIGH-10: agentSecret required in computeHashes/generateProof (no fallback)
- HIGH-13: ComputeBudgetProgram CU limits on all SDK tx functions
- HIGH-17: formatSol integer math (Math.trunc + modulo, no float)
- HIGH-18: NullifierCache with timestamp tracking + confirmUsed()

All 234 LiteSVM integration tests, 100 Rust unit tests, 257 SDK tests,
and 4689 runtime tests pass.

+++ programs/agenc-coordination/src/instructions/claim_task.rs
@@ -150,7 +150,7 @@ pub fn handler(ctx: Context<ClaimTask>) -> Result<()> {
     }
 
     // Check worker doesn't have too many active tasks
-    const MAX_ACTIVE_TASKS: u8 = 10;
+    const MAX_ACTIVE_TASKS: u16 = 10;
     require!(
         worker.active_tasks < MAX_ACTIVE_TASKS,
         CoordinationError::MaxActiveTasksReached
